Create elderly_care
it is an elderly caregiver matchmaking project where the elderly individuals and their family will be able to find caregivers according to their preferences. There is an emergency support android application which can be used by elderly in times of emergency, additionaly a ML prediction module to predict stroke and cvd.